Add java-salary project
In this project, concepts of inversion of control and dependency injection were worked on and, as a result of this effort, it is possible to change, in the main program, the instance of components for the key component such as SalaryService without having to change it. The phenomenon of low coupling is observed in this project.